For Midterm

+++ location.management/src/main/java/kz/iitu/location/management/entity/Location.java
@@ -11,9 +11,9 @@ import javax.persistence.Table;
 public class Location {
     @Id
     @Column(name = "id")
-    private int id;
+    private Long id;
     @Column(name = "trip_id")
-    private int tripId;
+    private Long tripId;
     @Column(name = "first_loc")
     private String firstLoc;
     @Column(name  = "last_loc")
@@ -24,7 +24,7 @@ public class Location {
     public Location() {
     }
 
-    public Location(int id, int tripId, String firstLoc, String lastLoc, int seq) {
+    public Location(Long id, Long tripId, String firstLoc, String lastLoc, int seq) {
         this.id = id;
         this.tripId = tripId;
         this.firstLoc = firstLoc;
@@ -32,19 +32,19 @@ public class Location {
         this.seq = seq;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public int getTripId() {
+    public Long getTripId() {
         return tripId;
     }
 
-    public void setTripId(int tripId) {
+    public void setTripId(Long tripId) {
         this.tripId = tripId;
     }
